Fixed small casting bug + CI

+++ crates/bevy_gizmos/src/gizmos.rs
@@ -659,7 +659,7 @@ impl Drop for Arc2dBuilder<'_, '_> {
             Some(segments) => segments,
             // Do a linear interpolation between 1 and `DEFAULT_CIRCLE_SEGMENTS`
             // using the arc angle as scalar.
-            None => ((self.arc_angle.abs() / TAU) * DEFAULT_CIRCLE_SEGMENTS).ceil() as usize,
+            None => ((self.arc_angle.abs() / TAU) * DEFAULT_CIRCLE_SEGMENTS as f32).ceil() as usize,
         };
 
         let positions = arc_inner(self.direction_angle, self.arc_angle, self.radius, segments)
